+++ stats/analyzer/src/main/java/ru/yandex/practicum/graduation/stats/analyzer/service/RecommendationService.java
@@ -6,9 +6,9 @@ import ru.practicum.stats.dto.dto.RecommendationEvent;
 import java.util.List;
 
 public interface RecommendationService {
-    List<RecommendationEvent> getRecommendationForUser(Long userId, int maxResults);
+    List<RecommendationEvent> getRecommendationForUser(Long userId, long maxResults);
 
-    List<RecommendationEvent> getSimilarEvents(Long eventId, Long userId, int maxResults);
+    List<RecommendationEvent> getSimilarEvents(Long eventId, Long userId, long maxResults);
 
     List<RecommendationEvent> getInteractionsCount(List<Long> eventIds);
 }
